Moves tests to /tests folder.  Adds more server/protest.js tests.
