+++ package.json
@@ -16,8 +16,8 @@
     "start": "node server/start.js",
     "start-watch": "nodemon server/start.js --watch server --watch db --watch index.js --watch package.json",
     "start-dev": "cross-env NODE_ENV=development npm run start-watch",
-    "test": "mocha --compilers js:babel-register --require babel-polyfill --watch-extensions js,jsx app/**/*.test.js app/**/*.test.jsx server/**/*.test.js db/**/*.test.js",
-    "coverage": "nyc mocha --compilers js:babel-register --require babel-polyfill --watch-extensions js,jsx app/**/*.test.js app/**/*.test.jsx server/**/*.test.js db/**/*.test.js",
+    "test": "mocha --compilers js:babel-register --require babel-polyfill --watch-extensions js,jsx tests/**/*.test.js tests/**/*.test.jsx",
+    "coverage": "nyc mocha --compilers js:babel-register --require babel-polyfill --watch-extensions js,jsx tests/**/*.test.js tests/**/*.test.jsx",
     "test-watch": "npm run test -- --watch --reporter=min",
     "seed": "node db/seed.js",
     "deploy-heroku": "bin/deploy-heroku.sh",
