+++ server/users.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const db = require('APP/db')
+const db = require('../db')
 const User = db.model('users')
 
 const {mustBeLoggedIn, forbidden} = require('./auth.filters')
