+++ tests/app/WhoAmI.test.jsx
@@ -6,7 +6,7 @@ import {spy} from 'sinon'
 chai.use(require('sinon-chai'))
 import {createStore} from 'redux'
 
-import WhoAmIContainer, {WhoAmI} from './WhoAmI'
+import WhoAmIContainer, {WhoAmI} from '../../app/components/WhoAmI'
 
 /* global describe it beforeEach */
 describe('<WhoAmI/>', () => {
