+++ tests/server/protests.test.js
@@ -1,13 +1,13 @@
 const request = require('supertest')
 const {expect} = require('chai')
-const db = require('APP/db'), {Protest} = db
-const app = require('./start')
+const db = require('../../db'), {Protest} = db
+const app = require('../../server/start')
 
 const protest = {
   text: 'This is a sample protest',
   color: 'rgb(0,0,0)'
 }
-
+const mockProtest = { id: null }
 /* global describe it before afterEach beforeEach */
 describe('/api/protests', () => {
   before('Await database sync', () => db.didSync)
@@ -17,6 +17,9 @@ describe('/api/protests', () => {
     Protest.create({
       text: protest.text,
       color: protest.color
+    })
+    .then(protest => {
+      mockProtest.id = protest.id
     })
   )
 
@@ -31,11 +34,14 @@ describe('/api/protests', () => {
 
   describe('GET /protests', () => {
     describe('gets all protests', () =>
-      it('responds with an empty object', () =>
+      it('responds with a single protest in an array', () =>
         request(app).get('/api/protests')
           .expect(200)
-          .then(res => expect(res.body.length).to.eql(1))
-      ))
+          .then(res => {
+            expect(res.body.length).to.eql(1)
+          })
+      )
+    )
   })
 
   describe('PUT /protests/upvote/:id (protest)', () => {
@@ -44,6 +50,14 @@ describe('/api/protests', () => {
         .put('/api/protests/upvote/0')
         .expect(400)
         .then(res => expect(res.body.error).to.eql('No protest found with matching id.'))
+      )
+  })
+
+  describe('PUT /protests/upvote/:id (protest)', () => {
+    it('200 when successful', () =>
+      request(app)
+        .put(`/api/protests/upvote/${mockProtest.id}`)
+        .expect(200)
       )
   })
 
@@ -56,12 +70,28 @@ describe('/api/protests', () => {
       )
   })
 
+  describe('PUT /protests/downvote/:id (protest)', () => {
+    it('200 when successful', () =>
+      request(app)
+        .put(`/api/protests/downvote/${mockProtest.id}`)
+        .expect(200)
+      )
+  })
+
   describe('DELETE /protests/:id (protest)', () => {
     it('error message generated when no protest found with matching id', () =>
       request(app)
         .delete('/api/protests/0')
         .expect(400)
         .then(res => expect(res.body.error).to.eql('No protest found with matching id.'))
+      )
+  })
+
+  describe('DELETE /protests/:id (protest)', () => {
+    it('202 when successful', () =>
+      request(app)
+        .delete(`/api/protests/${mockProtest.id}`)
+        .expect(202)
       )
   })
 })
